Consolidate client integration tests
